1. Добавлен Spring Security с привязкой к БД. 2. Удаление и редактирование объявлений только созданных user'ом или admin. 3. Тесты по объявлениям

+++ src/main/java/ru/skypro/homework/service/impl/AdsServiceImpl.java
@@ -2,6 +2,7 @@ package ru.skypro.homework.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import ru.skypro.homework.dto.Ads;
@@ -9,6 +10,7 @@ import ru.skypro.homework.dto.CreateAds;
 import ru.skypro.homework.dto.FullAds;
 import ru.skypro.homework.dto.ResponseWrapperAds;
 import ru.skypro.homework.exception.AdsNotFoundException;
+import ru.skypro.homework.exception.UserForbiddenException;
 import ru.skypro.homework.exception.UserNotRegisterException;
 import ru.skypro.homework.mapper.AdsMapper;
 import ru.skypro.homework.model.AdsEntity;
@@ -26,6 +28,9 @@ public class AdsServiceImpl implements AdsService {
     private final AdsRepository adsRepository;
 
     private final UserRepository userRepository;
+
+    private final UserValidatePermission validatePermission;
+
     @Override
     public ru.skypro.homework.dto.Ads addAds(CreateAds properties, MultipartFile image, Authentication authentication) {
         // mapping from dto to entity
@@ -38,18 +43,30 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public void deleteAds(Integer id) {
+    public void deleteAds(Integer id, Authentication authentication) {
+        UserEntity user = userRepository.findUserEntityByEmail(authentication.getName())
+                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + authentication.getName()));
         AdsEntity adsEntity = adsRepository.findById(id).orElseThrow(() -> new AdsNotFoundException(id));
-        adsRepository.deleteById(id);
+        if (validatePermission.isAdmin(user) || validatePermission.isAdsOwner(user, adsEntity)) {
+            adsRepository.deleteById(id);
+        } else {
+            throw new UserForbiddenException(user.getId());
+        }
     }
 
     @Override
     public Ads updateAds(Integer id, CreateAds createAds, Authentication authentication) {
+        UserEntity user = userRepository.findUserEntityByEmail(authentication.getName())
+                .orElseThrow(() -> new UsernameNotFoundException("User not found: " + authentication.getName()));
         AdsEntity adsEntity = adsRepository.findById(id).orElseThrow(() -> new AdsNotFoundException(id));
-        adsEntity.setDescription(createAds.getDescription());
-        adsEntity.setPrice(createAds.getPrice());
-        adsEntity.setTitle(createAds.getTitle());
-        return adsMapper.adsEntityToAds(adsRepository.save(adsEntity));
+
+        if (validatePermission.isAdmin(user) || validatePermission.isAdsOwner(user, adsEntity)) {
+            adsEntity.setDescription(createAds.getDescription());
+            adsEntity.setPrice(createAds.getPrice());
+            adsEntity.setTitle(createAds.getTitle());
+            return adsMapper.adsEntityToAds(adsRepository.save(adsEntity));
+        }
+        throw new UserForbiddenException(user.getId());
     }
 
     @Override
